Refactoring prototype
